requirements.txt
flask
openai>=1.0.0
google-generativeai
python-decouple
markdown
bleach
pymongo[srv]
boto3
python-jose[cryptography]
requests
mysql-connector-python
playwright
apscheduler
beautifulsoup4
pdfplumber
python-magic
pytesseract
numpy>=2.3.0
opencv-python-headless
PyPDF2
Pillow